Unified provider-agnostic storage web worker
Replace the SQLiteProvider-specific worker architecture with a single
unified worker that supports both SQLite and IDB backends:

- Create SQLiteProvider/index.web.ts: StorageProvider implementation for
  SQLite WASM, reshaped from the old worker.ts handle* functions
- Create lib/storage/worker.ts: provider-agnostic worker that dynamically
  imports the appropriate StorageProvider based on init message
- Create WorkerStorageProvider.ts: generic main-thread proxy replacing
  both SQLiteProvider/index.ts and IDBKeyValProvider web usage
- Update platforms/index.ts: use WorkerStorageProvider with backend
  selection based on OPFS availability
- Remove old SQLiteProvider/index.ts (web proxy) and worker.ts

Both IDB and SQLite now run off the main thread in the same unified
worker, with BroadcastChannel cross-tab sync handled at the worker level.
